refactor:  create common component for network policy form
